perf: extract out helpers from pages to fix bunele issue
