+++ src/__tests__/pages/teachers/curriculum/[subjectPhaseSlug]/[tab].test.tsx
@@ -6,13 +6,7 @@ import curriculumApi from "@/node-lib/curriculum-api-2023";
 import CurriculumInfoPage, {
   getStaticProps,
   getStaticPaths,
-  formatCurriculumUnitsData,
-  createThreadOptions,
-  createYearOptions,
-  createInitialYearFilterSelection,
-  createUnitsListingByYear,
 } from "@/pages/teachers/curriculum/[subjectPhaseSlug]/[tab]";
-import { fetchSubjectPhasePickerData } from "@/pages/teachers/curriculum";
 import {
   curriculumOverviewCMSFixture,
   curriculumOverviewMVFixture,
@@ -23,6 +17,14 @@ import subjectPhaseOptions from "@/browser-lib/fixtures/subjectPhaseOptions";
 import { mockPrerelease } from "@/utils/mocks";
 import { parseSubjectPhaseSlug } from "@/utils/curriculum/slugs";
 import "@/__tests__/__helpers__/ResizeObserverMock";
+import {
+  createInitialYearFilterSelection,
+  createThreadOptions,
+  createUnitsListingByYear,
+  createYearOptions,
+  fetchSubjectPhasePickerData,
+  formatCurriculumUnitsData,
+} from "@/pages-helpers/curriculum/docx/tab-helpers";
 
 const render = renderWithProviders();
