+++ src/components/CurriculumComponents/UnitsTab/UnitsTab.test.tsx
@@ -3,10 +3,10 @@ import { waitFor } from "@testing-library/react";
 
 import UnitsTab from "./UnitsTab";
 
-import { formatCurriculumUnitsData } from "@/pages/teachers/curriculum/[subjectPhaseSlug]/[tab]";
 import curriculumUnitsTabFixture from "@/node-lib/curriculum-api-2023/fixtures/curriculumUnits.fixture";
 import renderWithProviders from "@/__tests__/__helpers__/renderWithProviders";
 import "@/__tests__/__helpers__/ResizeObserverMock";
+import { formatCurriculumUnitsData } from "@/pages-helpers/curriculum/docx/tab-helpers";
 
 const render = renderWithProviders();
 const curriculumThreadHighlighted = jest.fn();
