+++ src/pages-helpers/curriculum/docx/builder/2_tableOfContents.ts
@@ -7,10 +7,10 @@ import {
   pointToDxa,
   wrapInLinkToBookmark,
 } from "../docx";
+import { formatCurriculumUnitsData } from "../tab-helpers";
 
 import { uncapitalizeSubject } from "./helper";
 
-import { formatCurriculumUnitsData } from "@/pages/teachers/curriculum/[subjectPhaseSlug]/[tab]";
 import { getYearGroupTitle } from "@/utils/curriculum/formatting";
 
 export default async function generate(
